# admin/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  baseModel.IdReq:
    properties:
      id:
        description: 数据ID
        example: 1
        type: integer
    type: object
  baseModel.PageRes:
    properties:
      list:
        description: 分页数据
      total:
        description: 数据总量
        example: 100
        type: integer
    type: object
...
  baseModel.ResBody:
    properties:
      code:
        description: 响应码
        example: S000/F000
        type: string
      data:
        description: 响应数据
      msg:
        description: 响应消息
        example: Success/Fail
        type: string
      unPop:
        description: 免弹窗提示
        example: true
        type: boolean
    type: object
  baseModel.SelectNumRes:
    properties:
      label:
        description: 展示名
        example: Name
        type: string
      value:
        description: 展示值
        example: 1
        type: integer
    type: object
...
  platModel.ResponseAddReq:
    properties:
      code:
        description: 响应码，仅示例，实际入库实时计算
        example: S101
        type: string
      enUs:
        description: 英文响应文言
        example: EnUs
        maxLength: 128
        type: string
      remark:
        description: 其他备注信息
        example: Remark
        maxLength: 128
        type: string
      serviceCode:
        description: 业务ID，来源于字典，指定响应码归属业务
        example: "0"
        maxLength: 3
        type: string
      type:
        description: 响应类型，该字段用于筛选，可配置S/F/E
        enum:
        - S
        - F
        - E
        example: S
        type: string
      zhCn:
        description: 中文响应文言
        example: ZhCn
        maxLength: 128
        type: string
    required:
    - enUs
    - serviceCode
    - type
    - zhCn
    type: object
  platModel.ResponseEditReq:
    properties:
      enUs:
        description: 英文响应文言
        example: EnUs
        maxLength: 128
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      remark:
        description: 其他备注信息
        example: Remark
        maxLength: 128
        type: string
      zhCn:
        description: 中文响应文言
        example: ZhCn
        maxLength: 128
        type: string
    required:
    - enUs
    - id
    - zhCn
    type: object
  platModel.ResponseGetRes:
    properties:
      code:
        description: 响应码
        example: "20101"
        type: string
      enUs:
        description: 英文响应文言
        example: EnUs
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      remark:
        description: 其他备注信息
        example: Remark
        type: string
      serviceCode:
        description: 业务ID，来源于字典，指定响应码归属业务
        example: "1"
        type: string
      type:
        description: 响应类型，该字段用于筛选，可配置S/F/E
        example: S
        type: string
      zhCn:
        description: 中文响应文言
        example: ZhCn
        type: string
    type: object
  platModel.ResponseNextValReq:
    properties:
      serviceCode:
        description: 业务ID，来源于字典，指定响应码归属业务
        example: "0"
        maxLength: 3
        type: string
      type:
        description: 响应类型，该字段用于筛选，可配置S/F/E
        enum:
        - S
        - F
        - E
        example: S
        type: string
    required:
    - serviceCode
    - type
    type: object
  platModel.ResponsePageReq:
    properties:
      code:
        description: 响应码，支持模糊查询
        example: "20"
        type: string
      current:
        description: 当前页
        example: 1
        type: integer
      pageSize:
        description: 单页数量
        example: 10
        type: integer
      serviceCode:
        description: 业务ID，来源于字典，指定响应码归属业务
        example: "0"
        type: string
      type:
        description: 响应类型，该字段用于筛选，可配置S/F/E
        example: S
        type: string
    type: object
  platModel.ResponsePageRes:
    properties:
      code:
        description: 响应码
        example: "20101"
        type: string
      enUs:
        description: 英文响应文言
        example: EnUs
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      mark:
        description: 变更标识 0可变更 1禁止变更
        example: "1"
        type: string
      remark:
        description: 其他备注信息
        example: Remark
        type: string
      serviceCode:
        description: 业务ID，来源于字典，指定响应码归属业务
        example: "1"
        type: string
      type:
        description: 响应类型，该字段用于筛选，可配置S/F/E
        example: S
        type: string
      zhCn:
        description: 中文响应文言
        example: ZhCn
        type: string
    type: object
  platModel.RoleAddReq:
    properties:
      halfPermissionIds:
        description: 权限集（父级半选）
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
      name:
        description: 角色名称
        example: Admin
        maxLength: 16
        type: string
      permissionIds:
        description: 权限集
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
      remark:
        description: 角色备注
        example: This is Admin
        maxLength: 64
        type: string
    required:
    - name
    type: object
  platModel.RoleEditReq:
    properties:
      halfPermissionIds:
        description: 权限集（父级半选）
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
      id:
        description: 数据ID
        example: 1
        type: integer
      name:
        description: 角色名称
        example: Admin
        maxLength: 16
        type: string
      permissionIds:
        description: 权限集
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
        uniqueItems: true
      remark:
        description: 角色备注
        example: This is Admin
        maxLength: 64
        type: string
    required:
    - id
    - name
    type: object
  platModel.RoleGetRes:
    properties:
      halfPermissionIds:
        description: 权限集（父级半选）
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      id:
        description: 数据ID
        example: 1
        type: integer
      name:
        description: 角色名称
        example: Admin
        type: string
      permissionIds:
        description: 权限集
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      remark:
        description: 角色备注
        example: This is Admin
        type: string
    type: object
  platModel.RolePageReq:
    properties:
      current:
        description: 当前页
        example: 1
        type: integer
      name:
        description: 角色名称，支持模糊搜索
        example: Admin
        type: string
      pageSize:
        description: 单页数量
        example: 10
        type: integer
    type: object
  platModel.RolePageRes:
    properties:
      halfPermissionIds:
        description: 权限集（父级半选）
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      id:
        description: 数据ID
        example: 1
        type: integer
      mark:
        description: 变更标识 0可变更 1禁止变更
        example: "1"
        type: string
      name:
        description: 角色名称
        example: Admin
        type: string
      permissionIds:
        description: 权限集
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      remark:
        description: 角色备注
        example: This is Admin
        type: string
    type: object
...
  platModel.SysConfigEditReq:
    properties:
      loginFailLockNum:
        description: 登陆失败锁定数量，最小为1
        example: 1
        type: integer
      loginFailLockUnit:
        description: 登陆失败锁定 1秒 2分 3时 4天
        example: "1"
        type: string
      loginFailNum:
        description: 登陆失败限制数，最小为1
        example: 1
        type: integer
      loginFailSwitch:
        description: 登陆失败限制开关，0限制 1不限制
        enum:
        - "0"
        - "1"
        example: "0"
        type: string
      loginFailTryNum:
        description: 登陆失败尝试次数
        example: 1
        type: integer
      loginFailUnit:
        description: 登陆失败限制 1秒 2分 3时 4天
        example: "1"
        type: string
      loginNum:
        description: 最大登陆并发量，最小为1
        example: 1
        type: integer
      loginSwitch:
        description: 并发限制开关，0限制 1不限制
        enum:
        - "0"
        - "1"
        example: "0"
        type: string
      logoutNum:
        description: 登陆过期长度数量，最小为1
        example: 1
        type: integer
      logoutSwitch:
        description: 登陆过期开关，0限制 1不限制
        enum:
        - "0"
        - "1"
        example: "0"
        type: string
      logoutUnit:
        description: 登陆过期单位，1秒 2分 3时 4天
        example: "1"
        type: string
    required:
    - loginFailSwitch
    - loginSwitch
    - logoutSwitch
    type: object
  platModel.SysConfigGetRes:
    properties:
      loginFailLockNum:
        description: 登陆失败锁定数量，最小为1
        example: 1
        type: integer
      loginFailLockUnit:
        description: 登陆失败锁定 1秒 2分 3时 4天
        example: "1"
        type: string
      loginFailNum:
        description: 登陆失败最大尝试次数，最小为1
        example: 1
        type: integer
      loginFailSwitch:
        description: 登陆失败限制开关，0限制 1不限制
        example: "0"
        type: string
      loginFailTryNum:
        description: 登陆失败尝试次数
        example: 1
        type: integer
      loginFailUnit:
        description: 登陆失败限制 1秒 2分 3时 4天
        example: "1"
        type: string
      loginNum:
        description: 最大登陆并发量，最小为1
        example: 1
        type: integer
      loginSwitch:
        description: 并发限制开关，0限制 1不限制
        example: "0"
        type: string
      logoutNum:
        description: 登陆过期长度数量，最小为1
        example: 1
        type: integer
      logoutSwitch:
        description: 登陆过期开关，0限制 1不限制
        example: "0"
        type: string
      logoutUnit:
        description: 登陆过期单位，1秒 2分 3时 4天
        example: "1"
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: dev@siteol.com
    name: smart
    url: https://smart.siteol.com
  description: "![Smart](http://localhost:8000/docs/file/logo.png)\r\n\r\n# 概述\r\n\r\n轻量易用的授权基座smart，提供一个非官方RBAC授权模型的基础开箱即用中台管理服务。\r\n\r\n支持响应码自定义国际化、高粒度权限分配、接口路由配置和实时生效，同时支持部门数据权限横向访问限制。\r\n\r\n本系统的全部接口采用【POST】【application/json】方式传输数据。\r\n\r\n除开放接口以外的其他接口均需要通过【ApiKeyAuth:请求头[Token]】完成鉴权。\r\n\r\n系统技术栈：Golang、VueNext、MySQL、Redis、Gin、ArcoDesign\r\n\r\n#
    接口文档\r\n\r\n轻度依赖集成，通过静态HTML加载yaml文件进行打开接口文档。\r\n\r\n该接口文档提供Swagger[支持调试]和ReDoc[阅读增强]两个版本。\r\n\r\n[Swagger[支持调试]：/docs/swagger/index.html](http://localhost:8000/docs/swagger/index.html)
    丨 [ReDoc[阅读增强]：/docs/redoc/index.html](http://localhost:8000/docs/redoc/index.html)\r\n\r\n##
    生成说明\r\n\r\n- 首次集成\r\n\r\ngo install github.com/swaggo/swag/cmd/swag@latest\r\n\r\n-
    格式化注释代码 (选用)\r\n\r\nswag fmt\r\n\r\n- 初始化swagger.yaml文件\r\n\r\nswag init\r\n\r\n-
    如果像本工程一样依赖了api.md和独立的api.go文件描述项目（推荐，后续持续使用）\r\n\r\nswag init -g api.go --md .\r\n\r\n-
    删除多余的生成\r\n\r\nrm .\\docs\\docs.go\r\n\r\n删除docs.go是因为本项目中并未采用下述依赖来集成。\r\n\r\n-
    github.com/swaggo/swag\r\n- github.com/swaggo/gin-swagger\r\n- github.com/swaggo/files\r\n\r\n而是通过HTML模板+JS+YAML引入集成，移除docs.go用于避免依赖编译报错。\r\n\r\n源帮助页：[https://github.com/swaggo/swag/blob/master/README_zh-CN.md](https://github.com/swaggo/swag/blob/master/README_zh-CN.md)\r\n"
  title: smart
  version: "1.0"
  x-logo:
    altText: smart
...
paths:
...
  /plat/response/add:
    post:
      consumes:
      - application/json
      description: 新建响应码配置
      operationId: AddResponse 响应码新建
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.ResponseAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 响应码新建
      tags:
      - 响应配置
  /plat/response/del:
    post:
      consumes:
      - application/json
      description: 响应码封存处理
      operationId: DelResponse 响应码封存
      parameters:
      - description: 语言，不传默认为zh-CN
        in: header
        name: Lang
        type: string
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          items:
            $ref: '#/definitions/baseModel.IdReq'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 响应码封存
      tags:
      - 响应配置
  /plat/response/edit:
    post:
      consumes:
      - application/json
      description: 在响应码分组下编辑响应码
      operationId: EditResponse 响应码编辑
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.ResponseEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 响应码编辑
      tags:
      - 响应配置
  /plat/response/get:
    post:
      consumes:
      - application/json
      description: 查询响应码详情
      operationId: GetResponse 响应码详情
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/platModel.ResponseGetRes'
              type: object
      security:
      - Token: []
      summary: 响应码详情
      tags:
      - 响应配置
  /plat/response/nextVal:
    post:
      consumes:
      - application/json
      description: 在业务分组下读取下一个响应码的建议值
      operationId: NextResponseVal 响应码NextVal建议
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.ResponseNextValReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - Token: []
      summary: 响应码NextVal建议
      tags:
      - 响应配置
  /plat/response/page:
    post:
      consumes:
      - application/json
      description: 查询响应码分页数据
      operationId: PageResponse 响应码分页
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.ResponsePageReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/baseModel.PageRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/platModel.ResponsePageRes'
                        type: array
                    type: object
              type: object
      security:
      - Token: []
      summary: 响应码分页
      tags:
      - 响应配置
  /plat/role/add:
    post:
      consumes:
      - application/json
      description: 新建角色配置
      operationId: AddRole 角色新建
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.RoleAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 角色新建
      tags:
      - 角色配置
  /plat/role/del:
    post:
      consumes:
      - application/json
      description: 角色删除处理
      operationId: DelRole 角色删除
      parameters:
      - description: 语言，不传默认为zh-CN
        in: header
        name: Lang
        type: string
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 角色删除
      tags:
      - 角色配置
  /plat/role/edit:
    post:
      consumes:
      - application/json
      description: 在角色分组下编辑角色
      operationId: EditRole 角色编辑
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.RoleEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 角色编辑
      tags:
      - 角色配置
  /plat/role/get:
    post:
      consumes:
      - application/json
      description: 查询角色详情
      operationId: GetRole 角色详情
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/platModel.RoleGetRes'
              type: object
      security:
      - Token: []
      summary: 角色详情
      tags:
      - 角色配置
  /plat/role/list:
    post:
      consumes:
      - application/json
      description: 角色下拉列表
      operationId: ListRole 角色下拉列表
      parameters:
      - description: 语言，不传默认为zh-CN
        in: header
        name: Lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baseModel.SelectNumRes'
                  type: array
              type: object
      security:
      - Token: []
      summary: 角色列表
      tags:
      - 角色配置
  /plat/role/page:
    post:
      consumes:
      - application/json
      description: 查询角色分页数据
      operationId: PageRole 角色分页
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.RolePageReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/baseModel.PageRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/platModel.RolePageRes'
                        type: array
                    type: object
              type: object
      security:
      - Token: []
      summary: 角色分页
      tags:
      - 角色配置
...
  /plat/sysConfig/edit:
    post:
      consumes:
      - application/json
      description: 在系统配置分组下编辑系统配置
      operationId: EditSysConfig 系统配置编辑
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.SysConfigEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 系统配置编辑
      tags:
      - 数据系统配置
  /plat/sysConfig/get:
    post:
      consumes:
      - application/json
      description: 查询系统配置详情
      operationId: GetSysConfig 系统配置详情
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/platModel.SysConfigGetRes'
              type: object
      security:
      - Token: []
      summary: 系统配置详情
      tags:
      - 数据系统配置
securityDefinitions:
  Token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"
...